backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
dnspython
PyMuPDF
python-docx
google-generativeai
openai
numpy
scikit-learn
pydantic
python-multipart
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]
email-validator
motor